@@ -1,42 +1,75 @@
-source 'http://rubygems.org'
+source 'https://rubygems.org'
 
-gem "rake", "0.8.7"
-gem "rails", "3.0.0"
-gem "ruby-mysql", "2.9.4"
-gem "ruby-ldap", "0.9.11"
-gem "will_paginate", "3.0.pre2"
-gem "jpmobile", "1.0.12"
-gem "net-ssh", "2.0.23"
-gem "rmagick", "2.13.1"
-gem "mail", "2.2.6.1"
-gem "hpricot", "0.8.3"
-gem "tamtam", "0.0.3"
-gem "zipruby", "0.3.6"
-gem "mime-types", "2.6.1"
 
-# Bundle edge Rails instead:
-# gem 'rails', :git => 'git://github.com/rails/rails.git'
+# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
+gem 'rails', '4.2.7'
+# Use mysql as the database for Active Record
+gem 'mysql2', '>= 0.3.13', '< 0.5'
+# Use SCSS for stylesheets
+gem 'sass-rails', '~> 5.0'
+# Use Uglifier as compressor for JavaScript assets
+gem 'uglifier', '>= 1.3.0'
+# Use CoffeeScript for .coffee assets and views
+gem 'coffee-rails', '~> 4.1.0'
+# See https://github.com/rails/execjs#readme for more supported runtimes
+gem 'therubyracer', platforms: :ruby
 
-# gem 'sqlite3-ruby', :require => 'sqlite3'
+# Use jquery as the JavaScript library
+#gem 'jquery-rails'
+# Turbolinks makes following links in your web application faster. Read more: https://github.com/rails/turbolinks
+#gem 'turbolinks'
+# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
+gem 'jbuilder', '~> 2.0'
+# bundle exec rake doc:rails generates the API under doc/api.
+gem 'sdoc', '~> 0.4.0', group: :doc
 
-# Use unicorn as the web server
+# Use ActiveModel has_secure_password
+# gem 'bcrypt', '~> 3.1.7'
+
+# Use Unicorn as the app server
 # gem 'unicorn'
 
-# Deploy with Capistrano
-# gem 'capistrano'
+# Use Capistrano for deployment
+# gem 'capistrano-rails', group: :development
 
-# To use debugger
-# gem 'ruby-debug'
+gem 'activerecord-import', '~> 0.15.0'
+gem 'activerecord-session_store', '~> 1.0.0'
+gem 'active_hash', '~> 1.4.1'
+gem 'enumerize', '~> 1.1.1'
+gem 'ruby-ldap', '~> 0.9.11'
+gem 'will_paginate', '~> 3.1.0'
+gem 'jpmobile', '~> 4.2.0'
+gem 'rmagick', '~> 2.15.4'
+gem 'nokogiri', '~> 1.6.8'
+gem 'nokogumbo', '~> 1.4.7'
+gem 'hpricot', '~> 0.8.6'
+gem 'premailer', '~> 1.8.7'
+gem 'sanitize', '~> 4.0.1'
+gem 'zipruby', '0.3.6'
+gem 'mime-types', '2.6.1'
+gem 'shared-mime-info', '~> 0.2.5'
+gem 'mail', '2.6.4.rc2'
+gem 'mail-iso-2022-jp', '2.0.5.rc1'
+gem 'charlock_holmes', '~> 0.7.3'
+gem 'delayed_job_active_record', '~> 4.1.1'
+gem 'daemons', '~> 1.2.3'
 
-# Bundle the extra gems:
-# gem 'bj'
-# gem 'nokogiri'
-# gem 'sqlite3-ruby', :require => 'sqlite3'
-# gem 'aws-s3', :require => 'aws/s3'
+gem 'dynamic_form', '~> 1.1.4'
 
-# Bundle gems for the local environment. Make sure to
-# put test-only gems in this group so their generators
-# and rake tasks are available in development mode:
-# group :development, :test do
-#   gem 'webrat'
-# end
+gem 'bower-rails', '~> 0.10.0'
+gem 'whenever', '~> 0.9.4', require: false
+
+group :development, :test do
+  # Call 'byebug' anywhere in the code to stop execution and get a debugger console
+  gem 'byebug'
+end
+
+group :development do
+  # Access an IRB console on exception pages or by using <%= console %> in views
+  gem 'web-console', '~> 2.0'
+
+  # Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
+  gem 'spring'
+
+  gem 'quiet_assets', '~> 1.1.0'
+end
