@@ -1,25 +1,22 @@
-# encoding: utf-8
 class Sys::Admin::AccountController < Sys::Controller::Admin::Base
-  protect_from_forgery :except => [:login]
-  
+  protect_from_forgery except: [:login]
+  layout 'base'
+
   def login
-    skip_layout
-    #admin_uri = '/_admin'
-    admin_uri = '/_admin/gw/webmail/INBOX/mails'
-    
+    admin_uri = '/webmail/INBOX/mails'
+
     #return redirect_to(admin_uri) if logged_in?
     if request.mobile? && params[:_session_id] == ''
       return redirect_to admin_uri
     end
-    
+
     @uri = params[:uri] || cookies[:sys_login_referrer] || admin_uri
     @uri = @uri.gsub(/^http:\/\/[^\/]+/, '')
     @uri = NKF::nkf('-w', @uri)
     return unless request.post?
-    
+
     if params[:password].to_s == 'p' + params[:account].to_s
-      cond = {:account => params[:account]}
-      if Sys::User.find(:first, :conditions => cond)
+      if Sys::User.where(account: params[:account]).first
         flash.now[:notice] = "初期パスワードではログインできません。<br />パスワードを変更してください。".html_safe
         respond_to do |format|
           format.html { render }
@@ -28,22 +25,22 @@ class Sys::Admin::AccountController < Sys::Controller::Admin::Base
         return true
       end
     end
-    
-    if request.mobile?
+
+    if request.mobile? || request.smart_phone?
       login_ok = new_login_mobile(params[:account], params[:password], params[:mobile_password])
     else
       login_ok = new_login(params[:account], params[:password])
     end
-    
+
     unless login_ok
-      flash.now[:notice] = "ユーザＩＤ・パスワードを正しく入力してください"
+      flash.now[:notice] = "ユーザーＩＤ・パスワードを正しく入力してください"
       respond_to do |format|
         format.html { render }
         format.xml  { render(:xml => '<errors />') }
       end
       return true
     end
-    
+
     if params[:remember_me] == "1"
       user = Sys::User.find(self.current_user.id)
       user.remember_me
@@ -52,10 +49,9 @@ class Sys::Admin::AccountController < Sys::Controller::Admin::Base
         :expires => user.remember_token_expires_at
       }
     end
-    
+
     cookies.delete :sys_login_referrer
-    Sys::Session.delete_past_sessions_at_random
-      
+
     respond_to do |format|
       format.html { redirect_to @uri }
       format.xml  { render(:xml => current_user.to_xml) }
@@ -69,26 +65,26 @@ class Sys::Admin::AccountController < Sys::Controller::Admin::Base
     end
     cookies.delete :auth_token
     reset_session
-    redirect_to('action' => 'login')
+    redirect_to action: :login
   end
-  
+
   def info
-    skip_layout
-    
     respond_to do |format|
       format.html { render }
       format.xml  { render :xml => Core.user.to_xml(:root => 'item', :include => :groups) }
     end
   end
-  
+
   def sso
-    skip_layout
-    
     params[:to] ||= 'gw'
     raise 'SSOの設定がありません。' unless config = Joruri.config.sso_settings[params[:to].to_sym]
-    
-    @uri = "#{config[:usessl] ? "https" : "http"}://#{config[:host]}:#{config[:port]}/"
-    
+
+    @uri = URI::HTTP.build(
+      scheme: config[:usessl] ? 'https' : 'http',
+      host: config[:host],
+      port: config[:port],
+    )
+
     require 'net/http'
     Net::HTTP.version_1_2
     http = Net::HTTP.new(config[:host], config[:port])
@@ -96,20 +92,26 @@ class Sys::Admin::AccountController < Sys::Controller::Admin::Base
       http.use_ssl = true
       http.verify_mode = OpenSSL::SSL::VERIFY_NONE
     end
-    
+
     http.start do |agent|
-      parameters = "account=#{Core.user.account}&password=#{CGI.escape(Core.user.password.to_s)}&mobile_password=#{CGI.escape(Core.user.mobile_password.to_s)}"
-      response = agent.post("/#{config[:path]}", parameters)
+      response = agent.post(config[:path], {
+        account: Core.user.account,
+        password: Core.user.password,
+        mobile_password: Core.user.mobile_password
+      }.to_query)
       @token = response.body =~ /^OK/i ? response.body.gsub(/^OK /i, '') : nil
     end
-    
-    return redirect_to @uri unless @token
-    
-    @uri << "#{config[:path]}"
+
+    return redirect_to @uri.to_s unless @token
+
+    @uri.path = config[:path]
     if request.get?
-      @uri << "?account=#{Core.user.account}&token=#{@token}"
-      @uri << "&path=#{CGI.escape(params[:path])}" if params[:path]
-      return redirect_to @uri
+      @uri.query = Hash.new.tap { |h|
+        h[:account] = Core.user.account
+        h[:token] = @token
+        h[:path] = params[:path] if params[:path].present?
+      }.to_query
+      return redirect_to @uri.to_s
     end
   end
 end
