@@ -1,12 +1,6 @@
-# encoding: utf-8
 class Sys::Admin::FrontController < Sys::Controller::Admin::Base
   def index
-    item = Sys::Message.new.public
-    @messages = item.find(:all, :order => 'published_at DESC')
-    
-    item = Sys::Maintenance.new.public
-    @maintenances = item.find(:all, :order => 'published_at DESC')
-    
-    #@calendar = Util::Date::Calendar.new(nil, nil)
+    @messages = Sys::Message.state_public.order(published_at: :desc)
+    @maintenances = Sys::Maintenance.state_public.order(published_at: :desc)
   end
 end
