@@ -1,39 +1,33 @@
-# encoding: utf-8
+require 'csv'
 class Sys::Admin::Groups::ExportController < Sys::Controller::Admin::Base
   include Sys::Controller::Scaffold::Base
-  require 'csv'
-  
+
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:manager)
   end
-  
+
   def index
-    
   end
- 
+
   def export
     if params[:do] == 'groups'
       export_groups
     elsif params[:do] == 'users'
       export_users
     else
-      return redirect_to(:action => :index)
+      return redirect_to(action: :index)
     end
   end
 
-  def all_groups(parent_id = 1)
-    groups = []
-    Sys::Group.find(:all, :conditions => {:parent_id => parent_id}, :order => :sort_no).each do |g|
-      groups << g
-      groups += all_groups(g.id)
-    end
-    groups
-  end
+  private
 
   def export_groups
-    csv = CSV.generate do |csv|
-      csv << [:code, :parent_code, :state, :web_state, :level_no, :sort_no,
-        :layout_id, :ldap, :ldap_version, :name, :name_en, :tel, :outline_uri, :email]
+    data = CSV.generate do |csv|
+      csv << [
+        :code, :parent_code, :state, :web_state, :level_no, :sort_no,
+        :layout_id, :ldap, :ldap_version, :name, :name_en, :tel, :outline_uri, :email
+      ]
+      all_groups = Sys::Group.roots.map(&:descendants).flatten.reject(&:root?)
       all_groups.each do |group|
         row = []
         row << group.code
@@ -53,15 +47,17 @@ class Sys::Admin::Groups::ExportController < Sys::Controller::Admin::Base
         csv << row
       end
     end
-    csv = NKF.nkf('-Ws -Lw', csv)
-    send_data(csv, :type => 'text/csv; charset=Shift_JIS', :filename => "sys_groups_#{Time.now.to_i}.csv")
+
+    data = NKF.nkf('-Ws -Lw', data)
+    send_data(data, type: 'text/csv; charset=Shift_JIS', filename: "sys_groups_#{Time.now.to_i}.csv")
   end
 
   def export_users
-    csv = CSV.generate do |csv|
-      csv << [:account, :state, :name, :name_en, :email, :auth_no, :password, :ldap, :ldap_version,
-        :group_code]
-      Sys::User.find(:all, :order => :id).each do |user|
+    data = CSV.generate do |csv|
+      csv << [
+        :account, :state, :name, :name_en, :email, :auth_no, :password, :ldap, :ldap_version, :group_code
+      ]
+      Sys::User.order(:id).each do |user|
         next unless user.groups[0]
         row = []
         row << user.account
@@ -77,7 +73,8 @@ class Sys::Admin::Groups::ExportController < Sys::Controller::Admin::Base
         csv << row
       end
     end
-    csv = NKF.nkf('-Ws -Lw', csv)
-    send_data(csv, :type => 'text/csv; charset=Shift_JIS', :filename => "sys_users_#{Time.now.to_i}.csv")
+
+    data = NKF.nkf('-Ws -Lw', data)
+    send_data(data, type: 'text/csv; charset=Shift_JIS', filename: "sys_users_#{Time.now.to_i}.csv")
   end
 end
