@@ -1,7 +1,6 @@
-# encoding: utf-8
+require 'csv'
 class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
   include Sys::Controller::Scaffold::Base
-  require 'csv'
 
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:manager)
@@ -12,34 +11,36 @@ class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
 
   def import
     if !params[:item] || !params[:item][:file]
-      return redirect_to(:action => :index)
+      return redirect_to(action: :index)
     end
 
     @results = [0, 0, 0]
 
-    require 'nkf'
     csv = NKF.nkf('-w', params[:item][:file].read)
 
+    messages = []
     if params[:do] == 'groups'
-      Core.messages << "インポート： グループ"
+      messages << "インポート： グループ"
       import_groups(csv)
     elsif params[:do] == 'users'
-      Core.messages << "インポート： ユーザ"
+      messages << "インポート： ユーザー"
       import_users(csv)
     else
-      return redirect_to(:action => :index)
+      return redirect_to(action: :index)
     end
 
-    Core.messages << "-- 追加 #{@results[0]}件"
-    Core.messages << "-- 更新 #{@results[1]}件"
-    Core.messages << "-- 失敗 #{@results[2]}件"
+    messages << "-- 追加 #{@results[0]}件"
+    messages << "-- 更新 #{@results[1]}件"
+    messages << "-- 失敗 #{@results[2]}件"
 
-    flash[:notice] = "インポートが終了しました。<br />#{Core.messages.join('<br />')}".html_safe
-    return redirect_to(:action => :index)
+    flash[:notice] = "インポートが終了しました。<br />#{messages.join('<br />')}".html_safe
+    return redirect_to(action: :index)
   end
 
+  private
+
   def import_groups(csv)
-    CSV.parse(csv, :headers => true, :header_converters => :symbol) do |data|
+    CSV.parse(csv, headers: true, header_converters: :symbol) do |data|
       code        = data[:code]
       parent_code = data[:parent_code]
 
@@ -48,13 +49,12 @@ class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
         next
       end
 
-      unless parent = Sys::Group.find_by_code(parent_code)
+      unless parent = Sys::Group.find_by(code: parent_code)
         @results[2] += 1
         next
       end
 
-      group = Sys::Group.find_by_code(code) || Sys::Group.new({:code => code})
-
+      group = Sys::Group.where(code: code).first_or_initialize
       group.parent_id    = parent.id
       group.state        = data[:state]
       group.web_state    = data[:web_state]
@@ -68,7 +68,7 @@ class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
       group.tel          = data[:tel]
       group.outline_uri  = data[:outline_uri]
       group.email        = data[:email]
-      
+
       next unless group.changed?
       status = group.new_record? ? 0 : 1
       if group.save
@@ -78,9 +78,9 @@ class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
       end
     end
   end
-  
+
   def import_users(csv)
-    CSV.parse(csv, :headers => true, :header_converters => :symbol) do |data|
+    CSV.parse(csv, headers: true, header_converters: :symbol) do |data|
       account     = data[:account]
       group_code  = data[:group_code]
 
@@ -89,12 +89,12 @@ class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
         next
       end
 
-      unless group = Sys::Group.find_by_code(group_code)
+      unless group = Sys::Group.find_by(code: group_code)
         @results[2] += 1
         next
       end
 
-      user = Sys::User.find_by_account(account) || Sys::User.new({:account => account})
+      user = Sys::User.where(account: account).first_or_initialize
       user.state        = data[:state]
       user.ldap         = data[:ldap]
       user.ldap_version = data[:ldap_version]
@@ -104,12 +104,12 @@ class Sys::Admin::Groups::ImportController < Sys::Controller::Admin::Base
       user.password     = data[:password]
       user.email        = data[:email]
       user.in_group_id  = group.id if group.id != user.group_id
-      
+
       next unless user.changed?
-      
+
       status = user.new_record? ? 0 : 1
       status = 2 unless user.save
-      
+
       @results[status] += 1
     end
   end
