@@ -1,11 +1,10 @@
-# encoding: utf-8
 class Sys::Admin::LdapGroupsController < Sys::Controller::Admin::Base
   include Sys::Controller::Scaffold::Base
   
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:manager)
-    return render(:text=> "LDAPサーバに接続できません。", :layout => true) unless Core.ldap.connection
-    
+    return render text: "LDAPサーバーに接続できません。", layout: true unless Core.ldap.connection
+
     if params[:parent] == '0'
       @parent  = nil
       @parents = []
@@ -14,7 +13,7 @@ class Sys::Admin::LdapGroupsController < Sys::Controller::Admin::Base
       @parents = @parent.parents
     end
   end
-  
+
   def index
     if !@parent
       @groups = Core.ldap.group.children
