@@ -1,14 +1,13 @@
-# encoding: utf-8
 module ApplicationHelper
   ## nl2br
   def br(str)
-    str.gsub(/\r\n|\r|\n/, '<br />')
+    str.gsub(/\r\n|\r|\n/, '<br />').html_safe
   end
   
   ## nl2br and escape
   def hbr(str)
     str = html_escape(str)
-    str.gsub(/\r\n|\r|\n/, '<br />')
+    str.gsub(/\r\n|\r|\n/, '<br />').html_safe
   end
   
   ## wrap long string
@@ -45,12 +44,12 @@ module ApplicationHelper
   def paginate(items, options = {})
     return '' unless items
     defaults = {
-      :params         => p,
-      :previous_label => '前のページ',
-      :next_label     => '次のページ',
-      :separator      => '<span class="separator"> | </span' + "\n" + '>'
+      params:         p,
+      previous_label: '前のページ',
+      next_label:     '次のページ',
+      link_separator: '<span class="separator"> | </span>'
     }
-    if request.mobile? && !request.smart_phone?
+    if request.mobile?
       defaults[:page_links]     = false
       defaults[:previous_label] = '&lt;&lt;*前へ'
       defaults[:next_label]     = '次へ#&gt;&gt;'
@@ -62,37 +61,20 @@ module ApplicationHelper
         m.gsub(/^(href=").*/, '\1' + URI.encode(Core.request_uri))
       end
     end
-    if request.mobile? && !request.smart_phone?
+    if request.mobile?
       links.gsub!(/<a [^>]*?rel="prev( |")/) {|m| m.gsub(/<a /, '<a accesskey="*" ')}
       links.gsub!(/<a [^>]*?rel="next( |")/) {|m| m.gsub(/<a /, '<a accesskey="#" ')}
     end
-    links
-  end
-  
-###  ## Emoji
-###  def emoji(name)
-###    require 'jpmobile'
-###    return Cms::Lib::Mobile::Emoji.convert(name, request.mobile)
-###  end
-###  
-###  ## Furigana
-###  def ruby(str, ruby = nil)
-###    ruby = Page.ruby unless ruby
-###    return ruby == true ? Cms::Lib::Navi::Ruby.convert(str) : str
-###  end
-  
-  ## Number format
-  def number_format(num)
-    number_to_currency(num, :unit => '', :precision => 0)
+    links.html_safe
   end
 
-  #show tag if condition is true.
-  def show_tag_if(tag, cond, options = {}, &block)
-    unless cond
-      options[:style] ||= ''
-      options[:style] += 'display:none;'
+  def content_tag_if(cond, tag, options = {}, &block)
+    if cond
+      content_tag tag, options do
+        capture(&block)
+      end
+    else
+      capture(&block)
     end
-    
-    content_tag(tag, options, &block)
   end
 end
