@@ -1,5 +1,3 @@
 class Sys::Sequence < ActiveRecord::Base
-  set_table_name "sys_sequences"
-  
-  scope :versioned, lambda{ |v| { :conditions => ["version = ?", "#{v}"] }}
+  scope :versioned, ->(v) { where(version: v) }
 end
