@@ -2,9 +2,7 @@ class Sys::UsersGroup < Sys::ManageDatabase
   include Sys::Model::Base
   include Sys::Model::Base::Config
   include Sys::Model::Auth::Manager
-  
-#  set_primary_key :rid
-  
-  belongs_to   :user,  :foreign_key => :user_id,  :class_name => 'Sys::User'
-  belongs_to   :group, :foreign_key => :group_id, :class_name => 'Sys::Group'
+
+  belongs_to :user, foreign_key: :user_id, class_name: 'Sys::User'
+  belongs_to :group, foreign_key: :group_id, class_name: 'Sys::Group'
 end
