@@ -1,57 +1,49 @@
-# encoding: utf-8
 class System::ProductSynchro < System::Database
   include Sys::Model::Base
   include Sys::Model::Base::Config
   include Sys::Model::Auth::Manager
-  
-  belongs_to :product, :class_name => 'System::Product', :foreign_key => :product_id
-  
+
+  belongs_to :product, class_name: 'System::Product', foreign_key: :product_id
+
   def execute
     # 中間データ作成
     return unless self.copy_ldap_temporary
-    
     # バックアップ
     return unless self.backup_table
-    
     # 同期
     self.synchronize
   end
-  
-protected
-  
+
+  private
+
   def copy_ldap_temporary
-    update_attributes(:state => 'temp')
+    update_attributes(state: 'temp')
     
-    @results = {:group => 0, :gerr => 0, :user => 0, :uerr => 0}
-    
-    item = System::LdapTemporary.new
-    item.and :version, version
-    item.and :parent_id, 0
-    item.and :data_type, 'group'
-    groups = item.find(:all, :order => 'sort_no, code')
-    
-    groups.each {|group| copy_ldap_temporaries(group, nil)}
-    
+    @results = { group: 0, gerr: 0, user: 0, uerr: 0 }
+
+    groups = System::LdapTemporary.where(version: version, parent_id: 0, data_type: 'group').order(:sort_no, :code)
+    groups.each { |group| copy_ldap_temporaries(group, nil) }
+
     messages = []
     messages << "グループ #{@results[:group]}件"
     messages << "-- 失敗 #{@results[:gerr]}件" if @results[:gerr] > 0
-    messages << "ユーザ #{@results[:user]}件"
+    messages << "ユーザー #{@results[:user]}件"
     messages << "-- 失敗 #{@results[:uerr]}件" if @results[:uerr] > 0
-    update_attributes(:remark_temp => messages.join("\n"))
-    
+    update_attributes(remark_temp: messages.join("\n"))
+
     if @results[:gerr] > 0 || @results[:uerr] > 0
-      update_attributes(:state => 'failure')
+      update_attributes(state: 'failure')
       return false
     end
-    
+
     return true
   end
-  
+
   def backup_table
-    update_attributes(:state => 'back')
-    
-    results = {:copy => 0, :cerr => 0}
-    
+    update_attributes(state: 'back')
+
+    results = { copy: 0, cerr: 0 }
+
     [:sys_users, :sys_groups, :sys_users_groups].each do |table|
       begin
         conn = ActiveRecord::Base.connection
@@ -63,45 +55,45 @@ protected
         results[:cerr] += 1
       end
     end
-    
+
     messages = []
     messages << "テーブル #{results[:copy]}件"
     messages << "--失敗 #{results[:cerr]}件" if results[:cerr] > 0
-    update_attributes(:remark_back => messages.join("\n"))
-    
+    update_attributes(remark_back: messages.join("\n"))
+
     if results[:cerr] > 0
-      update_attributes(:state => 'failure')
+      update_attributes(state: 'failure')
       return false
     end
-    
+
     return true
   end
-  
+
   def synchronize
-    update_attributes(:state => 'sync')
-    
+    update_attributes(state: 'sync')
+
     results = Sys::LdapSynchro.synchronize(version)
-    
+
     messages = []
     messages << "グループ"
     messages << "-- 更新 #{results[:group]}件"
     messages << "-- 削除 #{results[:gdel]}件" if results[:gdel] > 0
     messages << "-- 失敗 #{results[:gerr]}件" if results[:gerr] > 0
-    messages << "ユーザ"
+    messages << "ユーザー"
     messages << "-- 更新 #{results[:user]}件"
     messages << "-- 削除 #{results[:udel]}件" if results[:udel] > 0
     messages << "-- 失敗 #{results[:uerr]}件" if results[:uerr] > 0
-    update_attributes(:remark_sync => messages.join("\n"))
-    
+    update_attributes(remark_sync: messages.join("\n"))
+
     if results[:gerr] > 0 || results[:uerr] > 0
-      update_attributes(:state => 'failure')
+      update_attributes(state: 'failure')
       return false
     else
-      update_attributes(:state => 'success')
+      update_attributes(state: 'success')
       return true
     end
   end
-  
+
   def copy_ldap_temporaries(group, parent)
     ## group
     sg                = Sys::LdapSynchro.new
@@ -117,8 +109,8 @@ protected
     sg.official_position = group.official_position
     sg.assigned_job   = group.assigned_job
     sg.group_s_name   = group.group_s_name
-    
-    if sg.save(:validate => false)
+
+    if sg.save(validate: false)
       @results[:group] += 1
     else
       @results[:gerr] += 1
@@ -149,6 +141,16 @@ protected
     end
 
     ## next
-    group.ldap_children.each {|g| copy_ldap_temporaries(g, sg)}
+    group.ldap_children.each { |g| copy_ldap_temporaries(g, sg) }
+  end
+
+  class << self
+    def check
+      item = System::ProductSynchro.joins(:product)
+        .where('system_products.product_type = ?', 'mail')
+        .where(state: 'start')
+        .order(:created_at).first
+      item.execute if item
+    end
   end
 end
