@@ -1,5 +1,3 @@
-# encoding: utf-8
-
 ### site_name = "ジョールリ市"
 
 load "#{Rails.root}/db/seed/base.rb"
@@ -78,16 +76,16 @@ u4 = create 2, '伊藤　勝'    , 'user9', 'user9', 'user9@demo.joruri.org' # �
 ## ---------------------------------------------------------
 ## sys/users_groups
 
-g = Sys::Group.find_by_name_en('hisyokohoka')
-Sys::UsersGroup.update_all({:group_id => g.id}, {:user_id => 1})
+g = Sys::Group.find_by(name_en: 'hisyokohoka')
+Sys::UsersGroup.where(user_id: 1).update_all(group_id: g.id)
 Sys::UsersGroup.create :user_id => 2, :group_id => g.id
 Sys::UsersGroup.create :user_id => 3, :group_id => g.id
 Sys::UsersGroup.create :user_id => 4, :group_id => g.id
-g = Sys::Group.find_by_name_en('jinjika')
+g = Sys::Group.find_by(name_en: 'jinjika')
 Sys::UsersGroup.create :user_id => 5 , :group_id => g.id
 Sys::UsersGroup.create :user_id => 6 , :group_id => g.id
 Sys::UsersGroup.create :user_id => 7 , :group_id => g.id
-g = Sys::Group.find_by_name_en('kikakuseisakuka')
+g = Sys::Group.find_by(name_en: 'kikakuseisakuka')
 Sys::UsersGroup.create :user_id => 8 , :group_id => g.id
 Sys::UsersGroup.create :user_id => 9 , :group_id => g.id
 Sys::UsersGroup.create :user_id => 10, :group_id => g.id
@@ -95,214 +93,10 @@ Sys::UsersGroup.create :user_id => 10, :group_id => g.id
 ## ---------------------------------------------------------
 ## current_user
 
-Core.user       = Sys::User.find_by_account('admin')
+Core.user       = Sys::User.find_by(account: 'admin')
 Core.user_group = Core.user.groups[0]
 
-## ---------------------------------------------------------
-## cms/sites
-
-### Cms::Site.update_all({:name => site_name})
-
-## ---------------------------------------------------------
-## cms/concepts
-#
-#def create_cms_concept(params)
-#  params[:site_id]   ||= 1
-#  params[:parent_id] ||= 1
-#  params[:state]     ||= 'public'
-#  if params[:parent_id] == 0
-#    params[:level_no] = 1
-#  else
-#    parent = Cms::Concept.find_by_id(params[:parent_id])
-#    params[:level_no] = parent.level_no + 1
-#  end
-#  Cms::Concept.create(params)
-#end
-#
-#c_site  = Cms::Concept.find(1)
-#c_site.name = site_name
-#c_site.save
-#
-#c_top   = create_cms_concept :sort_no => 10 , :name => 'トップページ'
-#c_mayor = create_cms_concept :sort_no => 30 , :name => '市長室'
-#c_life  = create_cms_concept :sort_no => 50 , :name => 'ライフイベント'
-#
-## ---------------------------------------------------------
-## cms/contents
-#
-#def create_cms_content(params)
-#  params[:site_id]    ||= 1
-#  params[:concept_id] ||= 1
-#  params[:state]      ||= 'public'
-#  Cms::Content.create(params)
-#end
-#
-## ---------------------------------------------------------
-## sys/roles
-#
-#r = Sys::RoleName.create :name => 'common', :title => '一般ユーザ'
-#Sys::ObjectPrivilege.create :role_id => r.id, :item_unid => c_site.unid, :action => 'read'
-#Sys::UsersRole.create :user_id => u2.id, :role_id => r.id
-#Sys::UsersRole.create :user_id => u3.id, :role_id => r.id
-#Sys::UsersRole.create :user_id => u4.id, :role_id => r.id
-#
-## ---------------------------------------------------------
-## cms/layouts
-#
-#def create_cms_layout(params)
-#  params[:site_id]     ||= 1
-#  params[:concept_id]  ||= 1
-#  params[:state]       ||= 'public'
-#  params[:head]        ||= file("layouts/#{params[:name]}/head")
-#  params[:body]        ||= file("layouts/#{params[:name]}/body")
-#  params[:mobile_head] ||= file("layouts/#{params[:name]}/m_head")
-#  params[:mobile_body] ||= file("layouts/#{params[:name]}/m_body")
-#  Cms::Layout.create(params)
-#end
-#
-#l_top      = create_cms_layout :concept_id => c_top.id  , :name => 'top'             , :title => 'トップページ'
-#l_map      = create_cms_layout :concept_id => c_site.id , :name => 'sitemap'         , :title => 'サイトマップ'
-#l_mayor    = create_cms_layout :concept_id => c_mayor.id, :name => 'mayor'           , :title => '市長の部屋'
-#l_life     = create_cms_layout :concept_id => c_life.id , :name => 'lifeevent'       , :title => 'ライフイベント'
-#l_life_top = create_cms_layout :concept_id => c_life.id , :name => 'lifeevent-top'   , :title => 'ライフイベントTOP'
-#l_page     = create_cms_layout :concept_id => c_site.id , :name => 'page'            , :title => '詳細ページ'
-#l_top_emg1 = create_cms_layout :concept_id => c_top.id  , :name => 'emergency-level1', :title => '大規模災害トップページ１'
-#l_top_emg2 = create_cms_layout :concept_id => c_top.id  , :name => 'emergency-level2', :title => '大規模災害トップページ２'
-#l_top_emg3 = create_cms_layout :concept_id => c_top.id  , :name => 'emergency-level3', :title => '大規模災害トップページ３'
-#
-## ---------------------------------------------------------
-## cms/pieces
-#
-#def create_cms_piece(params)
-#  params[:site_id]        ||= 1
-#  params[:concept_id]     ||= 1
-#  params[:state]          ||= 'public'
-#  params[:body]           ||= file("pieces/#{params[:name]}/body")
-#  params[:xml_properties] ||= file("pieces/#{params[:name]}/xml_properties")
-#  Cms::Piece.create(params)
-#end
-#
-#[ [ c_site.id , 'Cms::Free'      , 'ad-lower'             , '広告（下部）' ],
-#  [ c_site.id , 'Cms::Free'      , 'ad-upper'             , '広告（右上部）' ],
-#  [ c_site.id , 'Cms::Free'      , 'address'              , '住所' ],
-#  [ c_site.id , 'Cms::PageTitle' , 'page-title'           , 'ページタイトル' ],
-#  [ c_site.id , 'Cms::BreadCrumb', 'bread-crumbs'         , 'パンくず' ],
-#  [ c_site.id , 'Cms::Free'      , 'global-navi'          , 'グローバルナビ' ],
-#  [ c_site.id , 'Cms::Free'      , 'footer-navi'          , 'フッターナビ' ],
-#  [ c_site.id , 'Cms::Free'      , 'common-banner'        , 'サイトバナー' ],
-#  [ c_site.id , 'Cms::Free'      , 'common-header'        , 'ふりがな・よみあげヘッダー' ],
-#  [ c_site.id , 'Cms::Free'      , 'recent-docs-title'    , '新着情報タイトル' ],
-#  [ c_site.id , 'Cms::Free'      , 'attract-information'  , '注目情報' ],
-#  [ c_site.id , 'Cms::Free'      , 'relation-link'        , '関連リンク' ],
-#  [ c_top.id  , 'Cms::Free'      , 'about'                , 'ジョールリ市の紹介' ],
-#  [ c_top.id  , 'Cms::Free'      , 'application'          , '申請書ダウンロード' ],
-#  [ c_top.id  , 'Cms::Free'      , 'area-information'     , '地域情報' ],
-#  [ c_top.id  , 'Cms::Free'      , 'basic-information'    , '基本情報' ],
-#  [ c_top.id  , 'Cms::Free'      , 'common-banner-top'    , 'サイトバナー（トップ）' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mayor'                , '市長室' ],
-#  [ c_top.id  , 'Cms::Free'      , 'qr-code'              , 'QRコード' ],
-#  [ c_top.id  , 'Cms::Free'      , 'photo'                , 'トップ写真' ],
-#  [ c_top.id  , 'Cms::Free'      , 'useful-information'   , 'お役立ち情報' ],
-#  [ c_top.id  , 'Cms::Free'      , 'topic'                , 'トピック' ],
-#  [ c_top.id  , 'Cms::Free'      , 'lifeevent'            , 'ライフイベント' ],
-#  [ c_top.id  , 'Cms::Free'      , 'category'             , 'カテゴリ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'inquiry'              , 'お問い合わせバナー' ],
-#  [ c_top.id  , 'Cms::Free'      , 'population'           , '人口' ],
-#  [ c_mayor.id, 'Cms::Free'      , 'mayor-side'           , '市長室サイドメニュー' ],
-#  [ c_mayor.id, 'Cms::Free'      , 'mayor'                , '市長室' ],
-#  [ c_mayor.id, 'Cms::Free'      , 'mayor-title'          , '市長室タイトル' ],
-#  [ c_life.id , 'Cms::Free'      , 'lifeevent-title'      , 'ライフイベントタイトル' ],
-#  [ c_life.id , 'Cms::Free'      , 'lifeevent-side'       , 'ライフイベントサイドメニュー' ],
-#  [ c_site.id , 'Cms::Free'      , 'mobile-common-header' , 'モバイル：ヘッダー画像' ],
-#  [ c_site.id , 'Cms::Free'      , 'mobile-copyright'     , 'モバイル：コピーライト' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-address'       , 'モバイル：住所' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-category-list' , 'モバイル：トップ分野一覧' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-footer-navi'   , 'モバイル：フッターナビ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-mayor'         , 'モバイル：ようこそ市長室へ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-menu-navi'     , 'モバイル：ナビ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-pickup'        , 'モバイル：ピックアップ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-recommend-site', 'モバイル：おすすめサイト' ],
-#  [ c_top.id  , 'Cms::Free'      , 'mobile-search'        , 'モバイル：サイト内検索' ],
-#  [ c_site.id , 'Cms::Free'      , 'mobile-back-navi'     , 'モバイル：バックナビ' ],
-#  [ c_site.id , 'Cms::Free'      , 'mobile-mayor-navi'    , 'モバイル：市長室' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency'                 , '大規模災害表示ピース　緊急アナウンス ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-application'     , '大規模災害表示ピース　申請書ダウンロード ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-area-information', '大規模災害表示ピース　地域情報 ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-common-banner'   , '大規模災害表示ピース　サイトバナー ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-info'            , '大規模災害表示ピース　緊急情報 ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-inquiry'         , '大規模災害表示ピース　お問い合わせバナー ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-link-bousai'     , '大規模災害表示ピース　リンク集(防災関係機関) ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-link-kasen'      , '大規模災害表示ピース　リンク集(河川情報) ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-link-koutsuu'    , '大規模災害表示ピース　リンク集(交通情報) ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-link-lifeline'   , '大規模災害表示ピース　リンク集(ライフライン情報) ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-mailmagazine'    , '大規模災害表示ピース　メールマガジン ' ],
-#  [ c_top.id  , 'Cms::Free'      , 'emergency-mode'            , '大規模災害表示ピース　モード表示' ],
-#].each do |c|
-#  create_cms_piece :concept_id => c[0], :model => c[1], :name => c[2], :title => c[3]
-#end
-#
-## ---------------------------------------------------------
-## cms/nodes
-#
-#def create_cms_node(params)
-#  params[:site_id]        ||= 1
-#  params[:concept_id]     ||= 1
-#  params[:parent_id]      ||= 1
-#  params[:state]          ||= 'public'
-#  params[:route_id]       ||= params[:parent_id]
-#  params[:directory]      ||= (params[:name] =~ /\./ ? 0 : 1)
-#  params[:published_at]   ||= Time.now
-#  Cms::Node.create(params)
-#end
-#
-#Cms::Node.update_all({:layout_id => l_top.id}, {:id => 1})
-#Cms::Node.update_all({:concept_id => c_top.id, :layout_id => l_top.id}, {:id => 2})
-#create_cms_node :parent_id => 1, :layout_id => l_page.id, :model => 'Cms::Page'     , :name => 'mobile.html', :title => 'ジョールリ市携帯サイトのご紹介', :body => file("nodes/pages/mobile/body")
-#
-#p = create_cms_node :layout_id => l_map.id, :model => 'Cms::Directory', :name => 'sitemap'   , :title => 'サイトマップ'
-#    create_cms_node :layout_id => l_map.id, :model => 'Cms::Sitemap'  , :name => 'index.html', :title => 'サイトマップ',
-#      :parent_id => p.id 
-#
-#m = create_cms_node :concept_id => c_mayor.id, :layout_id => l_mayor.id, :model => 'Cms::Directory', :name => 'mayor'     , :title => '市長室'
-#    create_cms_node :concept_id => c_mayor.id, :layout_id => l_mayor.id, :model => 'Cms::Page'     , :name => 'index.html', :title => '市長のご挨拶', :body => file("nodes/mayor/index/body"),
-#      :parent_id => m.id 
-#p = create_cms_node :parent_id => m.id, :concept_id => c_mayor.id, :layout_id => l_mayor.id, :model => 'Cms::Directory', :name => 'profile'   , :title => 'プロフィール'
-#    create_cms_node :parent_id => p.id, :concept_id => c_mayor.id, :layout_id => l_mayor.id, :model => 'Cms::Page'     , :name => 'index.html', :title => 'プロフィール', :body => file("nodes/mayor/dummy/body")
-#p = create_cms_node :parent_id => m.id, :concept_id => c_mayor.id, :layout_id => l_mayor.id, :model => 'Cms::Directory', :name => 'activity'  , :title => '市長へのメール'
-#    create_cms_node :parent_id => p.id, :concept_id => c_mayor.id, :layout_id => l_mayor.id, :model => 'Cms::Page'     , :name => 'index.html', :title => '市長へのメール', :body => file("nodes/mayor/dummy/body")
-#
-#p = create_cms_node :parent_id => 1   , :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Directory', :name => 'lifeevent'    , :title => 'ライフイベント'
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life_top.id, :model => 'Cms::Page'     , :name => 'index.html'   , :title => 'ライフイベント', :body => file("nodes/life/index/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'fukushi.html' , :title => '福祉・介護'    , :body => file("nodes/life/fukushi/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'hikkoshi.html', :title => '引越し'        , :body => file("nodes/life/hikkoshi/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'kekkon.html'  , :title => '結婚・離婚'    , :body => file("nodes/life/kekkon/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'kosodate.html', :title => '子育て・教育'  , :body => file("nodes/life/kosodate/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'ninshin.html' , :title => '妊娠・出産'    , :body => file("nodes/life/ninshin/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'seijin.html'  , :title => '成人になったら', :body => file("nodes/life/seijin/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'shibo.html'   , :title => '死亡'          , :body => file("nodes/life/shibo/body")
-#    create_cms_node :parent_id => p.id, :concept_id => c_life.id, :layout_id => l_life.id    , :model => 'Cms::Page'     , :name => 'shushoku.html', :title => '就職・退職'    , :body => file("nodes/life/shushoku/body")
-#
-## ---------------------------------------------------------
-## cms/site_settings
-#
-#def create_cms_site_setting(params)
-#  params[:site_id]        ||= 1
-#  Cms::SiteSetting.create(params)
-#end
-#
-#create_cms_site_setting(:name => "emergency_layout", :value => l_top.id     , :sort_no => 0)
-#create_cms_site_setting(:name => "emergency_layout", :value => l_top_emg1.id, :sort_no => 10)
-#create_cms_site_setting(:name => "emergency_layout", :value => l_top_emg2.id, :sort_no => 20)
-#create_cms_site_setting(:name => "emergency_layout", :value => l_top_emg3.id, :sort_no => 30)
-#
-## ---------------------------------------------------------
-## other modules
-#
-#load_seed_file "demo/article.rb"
-#load_seed_file "demo/enquete.rb"
-#load_seed_file "demo/portal.rb"
-#load_seed_file "demo/newsletter.rb"
 load_seed_file "demo/sys.rb"
-load_seed_file "demo/gw.rb"
+load_seed_file "demo/webmail.rb"
 
 puts "Imported demo data."
