@@ -1,14 +1,12 @@
-# encoding: utf-8
 class Sys::Controller::Admin::Base < ApplicationController
   include Sys::Controller::Admin::Auth
-  helper Sys::FormHelper
-  rescue_from ActiveRecord::RecordNotFound, :with => :error_auth
-  before_filter :pre_dispatch
+  rescue_from ActiveRecord::RecordNotFound, with: :error_auth
+  before_action :pre_dispatch
   layout  'admin/sys'
-  
+
   def initialize_application
     return false unless super
-    
+
     @@current_user = false
     if authenticate
       Core.user          = current_user
@@ -18,25 +16,18 @@ class Sys::Controller::Admin::Base < ApplicationController
     end
     return false
   end
-  
+
   def pre_dispatch
     ## each processes before dispatch
   end
-  
-  def self.simple_layout
-    self.layout 'admin/base'
-  end
-  
-  def simple_layout
-    self.class.layout 'admin/base'
-  end
-  
-private
+
+  private
+
   def authenticate
     return true  if logged_in?
     return false if request.env['PATH_INFO'] =~ /^\/_admin\/login/
     uri  = request.env['PATH_INFO']
-    uri += "?#{request.env['QUERY_STRING']}" if !request.env['QUERY_STRING'].blank?
+    uri += "?#{request.env['QUERY_STRING']}" if request.env['QUERY_STRING'].present?
     cookies[:sys_login_referrer] = uri
     respond_to do |format|
       format.html { redirect_to('/_admin/login') }
@@ -44,7 +35,7 @@ private
     end
     return false
   end
-  
+
   def error_auth
     http_error 500, '権限がありません。'
   end
