@@ -1,55 +1,55 @@
-# encoding: utf-8
 module Sys::Model::Base::File
-  def self.included(mod)
-    mod.validates_presence_of :file, :if => "@_skip_upload != true"
-    mod.validates_presence_of :name, :title
-    mod.validate :validate_file_name
-    mod.validate :validate_upload_file
-    mod.after_save :upload_internal_file
-    mod.after_destroy :remove_internal_file
+  extend ActiveSupport::Concern
+
+  included do
+    validates :file, presence: true, if: "@_skip_upload != true"
+    validates :name, :title, presence: true
+    validate :validate_file_name
+    validate :validate_upload_file
+    after_save :upload_internal_file
+    after_destroy :remove_internal_file
   end
-  
+
   @@_maxsize = 50# MegaMytes
-  
+
   attr_accessor :file, :file_maxsize
-  
+
   def skip_upload(bool = true)
     @_skip_upload = bool
   end
-  
+
   def validate_file_name
     return true if name.blank?
-    
+
     if self.name !~ /^[0-9a-zA-Z\-\_\.]+$/
       errors.add :name, "は半角英数字を入力してください。"
     elsif self.name.count('.') != 1
-      errors.add(:name, 'を正しく入力してください。＜ファイル名.拡張子＞')
+      errors.add :name, 'を正しく入力してください。＜ファイル名.拡張子＞'
     elsif duplicated?
       errors.add :name, "は既に存在しています。"
       return false
     end
     self.title = self.name if title.blank?
   end
-  
+
   def validate_upload_file
     return true if file.blank?
-    
+
     maxsize = @maxsize || file_maxsize || @@_maxsize
     if file.size > maxsize.to_i  * (1024**2)
       errors.add :file, "が容量制限を超えています。＜#{maxsize}MB＞"
       return true
     end
-    
+
     self.mime_type    = file.content_type
     self.size         = file.size
     self.image_is     = 2
     self.image_width  = nil
     self.image_height = nil
-    
+
     @_file_data = file.read
-    
+
     begin
-      require 'RMagick'
       image = Magick::Image.from_blob(@_file_data).shift
       if image.format =~ /(GIF|JPEG|PNG)/
         self.image_is = 1
@@ -61,14 +61,14 @@ module Sys::Model::Base::File
     rescue NoMethodError
     end
   end
-  
+
   def upload_path
     md_dir  = "#{self.class.to_s.underscore.pluralize}"
     id_dir  = format('%08d', id).gsub(/(.*)(..)(..)(..)$/, '\1/\2/\3/\4/\1\2\3\4')
     id_file = format('%07d', id) + '.dat'
     "#{Rails.root}/upload/#{md_dir}/#{id_dir}/#{id_file}"
   end
-  
+
   def readable
     return self
   end
@@ -80,48 +80,48 @@ module Sys::Model::Base::File
   def deletable
     return self
   end
-  
+
   def readable?
     return true
   end
-  
+
   def creatable?
     return true
   end
-  
+
   def editable?
     return true
   end
-  
+
   def deletable?
     return true
   end
-  
+
   def image_file?
     image_is == 1 ? true : nil 
   end
-  
+
   def escaped_name
     CGI::escape(name)
   end
-  
+
   def united_name
     title + '(' + eng_unit + ')'
   end
-  
+
   def alt
     title.blank? ? name : title
   end
-  
+
   def image_size
     return '' unless image_file?
     "( #{image_width}x#{image_height} )"
   end
-  
+
   def duplicated?
     nil
   end
-  
+
   def css_class
     if ext = File::extname(name).downcase[1..5]
       conv = {
@@ -133,7 +133,7 @@ module Sys::Model::Base::File
     end
     return 'iconFile'
   end
-  
+
   def eng_unit
     _size = size
     return '' unless _size.to_s =~ /^[0-9]+$/
@@ -150,7 +150,7 @@ module Sys::Model::Base::File
       _kilo = 0
       _unit = ''
     end
-    
+
     if _kilo > 0
       _size = (_size.to_f / (1024**_kilo)).to_s + '000'
       _keta = _size.index('.')
@@ -161,13 +161,13 @@ module Sys::Model::Base::File
         _size = _size.to_f.ceil.to_f / (10**(3-_keta))
       end
     end
-    
+
     "#{_size}#{_unit}Bytes"
   end
-  
+
   def reduced_size(options = {})
     return nil unless image_file?
-    
+
     src_w  = image_width.to_f
     src_h  = image_height.to_f
     dst_w  = options[:width].to_f
@@ -183,21 +183,20 @@ module Sys::Model::Base::File
     if options[:css]
       return "width: #{dst_w.ceil}px; height:#{dst_h.ceil}px;"
     end
-    return {:width => dst_w.ceil, :height => dst_h.ceil}
+    return { width: dst_w.ceil, height: dst_h.ceil }
   end
-  
+
   def mobile_image(mobile)
     return nil unless mobile
     return nil if image_is != 1
     return nil if image_width <= 300 && image_height <= 400
-    
+
     begin
-      require 'RMagick'
       #info = Magick::Image::Info.new
-      size = reduced_size(:width => 300, :height => 400)
+      size = reduced_size(width: 300, height: 400)
       img  = Magick::Image.read(public_path).first
       img  = img.resize(size[:width], size[:height])
-      
+
       case mobile
       when Jpmobile::Mobile::Docomo
         img.format = 'JPEG' if img.format == 'PNG'
@@ -212,16 +211,17 @@ module Sys::Model::Base::File
     end
     return img
   end
-  
-private
+
+  private
+
   ## filter/aftar_save
   def upload_internal_file
     unless @_file_data.blank?
-      Util::File.put(upload_path, :data => @_file_data, :mkdir => true)
+      Util::File.put(upload_path, data: @_file_data, mkdir: true)
     end
     return true
   end
-  
+
   ## filter/aftar_destroy
   def remove_internal_file
     return true unless FileTest.exist?(upload_path)
