@@ -1,45 +1,45 @@
-# encoding: utf-8
 class Sys::Admin::LanguagesController < Sys::Controller::Admin::Base
   include Sys::Controller::Scaffold::Base
-  
+
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:manager)
   end
-  
+
   def index
-    item = Sys::Language.new#.readable
-    item.page  params[:page], params[:limit]
-    item.order :sort_no
-    @items = item.find(:all)
+    @items = Sys::Language.order(:sort_no).paginate(page: params[:page], per_page: params[:limit])
     _index @items
   end
-  
+
   def show
-    @item = Sys::Language.new.find(params[:id])
+    @item = Sys::Language.find(params[:id])
     #return error_auth unless @item.readable?
-    
+
     _show @item
   end
 
   def new
-    @item = Sys::Language.new({
-      :state      => 'enabled',
-    })
+    @item = Sys::Language.new(state: 'enabled')
   end
-  
+
   def create
-    @item = Sys::Language.new(params[:item])
+    @item = Sys::Language.new(item_params)
     _create @item
   end
-  
+
   def update
-    @item = Sys::Language.new.find(params[:id])
-    @item.attributes = params[:item]
+    @item = Sys::Language.find(params[:id])
+    @item.attributes = item_params
     _update @item
   end
-  
+
   def destroy
-    @item = Sys::Language.new.find(params[:id])
+    @item = Sys::Language.find(params[:id])
     _destroy @item
+  end
+
+private
+
+  def item_params
+    params.require(:item).permit(:state, :name, :title, :sort_no)
   end
 end
