@@ -1,46 +1,29 @@
-require 'digest/md5'
 class Sys::File < ActiveRecord::Base
   include Sys::Model::Base
   include Sys::Model::Base::File
-  include Sys::Model::Rel::Unid
-  include Sys::Model::Rel::Creator
-  
-  #validates_presence_of :name
-  
-  ## garbage collect
-  def self.garbage_collect
-    conditions = Condition.new
-    conditions.and :tmp_id, 'IS NOT', nil
-    conditions.and :parent_unid, 'IS', nil
-    conditions.and :created_at, '<', (Date.strptime(Core.now, "%Y-%m-%d") - 2)
-    destroy_all(conditions.where)
-  end
-  
-  ## Remove the temporary flag.
-  def self.fix_tmp_files(tmp_id, parent_unid)
-    updates = {:parent_unid => parent_unid, :tmp_id => nil }
-    conditions = ["parent_unid IS NULL AND tmp_id = ?", tmp_id]
-    update_all(updates, conditions)
-  end
-  
-  def self.new_tmp_id
-    connection.execute("INSERT INTO #{table_name} (id, tmp_id) VALUES (null, 0)")
-    id = find_by_sql("SELECT LAST_INSERT_ID() AS id")[0].id
-    connection.execute("DELETE FROM #{table_name} WHERE id = #{id}")
-    Digest::MD5.new.update(id.to_s)
-  end
-  
+
   def duplicated?
-    file = self.class.new
-    file.and :id, "!=", id if id
-    file.and :name, name
+    rel = self.class.where(name: name)
+    rel = rel.where.not(id: id) if id
     if tmp_id
-      file.and :tmp_id, tmp_id
-      file.and :parent_unid, 'IS', nil
+      rel = rel.where(tmp_id: tmp_id)
+      rel = rel.where(parent_unid: nil)
     else
-      file.and :tmp_id, 'IS', nil
-      file.and :parent_unid, parent_unid
+      rel = rel.where(tmp_id: nil)
+      rel = rel.where(parent_unid: parent_unid)
     end
-    return file.find(:first)
+    rel.first
+  end
+
+  class << self
+    def new_tmp_id
+      SecureRandom.hex(16)
+    end
+
+    def garbage_collect
+      self.where.not(tmp_id: nil).where(parent_unid: nil)
+        .where("created_at < ?", 2.days.ago)
+        .find_each(&:destroy)
+    end
   end
 end
