@@ -1,21 +1,15 @@
 class Sys::Maintenance < ActiveRecord::Base
   include Sys::Model::Base
   include Sys::Model::Base::Page
-  include Sys::Model::Rel::Unid
-  include Sys::Model::Rel::Creator
   include Sys::Model::Auth::Manager
-  
-  belongs_to :status,  :foreign_key => :state, :class_name => 'Sys::Base::Status'
-  
-#  validate do |m|
-#    if published_at.blank?
-#      errors.add :in_published_at, :blank
-#    end
-#  end
 
-  #validates_presence_of :state, :published_at, :title, :body
-  validates_presence_of :state, :title, :body, :published_at
-    
-#  attr_accessor :in_published_at
-    
+  belongs_to_active_hash :status, foreign_key: :state, class_name: 'Sys::Base::Status'
+
+  validates :state, :title, :body, :published_at, presence: true
+
+  class << self
+    def latest_maintenance
+      self.state_public.order(published_at: :desc).first
+    end
+  end
 end
