@@ -1,68 +1,45 @@
-Joruri::Application.configure do
-  # Settings specified here will take precedence over those in config/environment.rb
+Rails.application.configure do
+  # Settings specified here will take precedence over those in config/application.rb.
 
   # The test environment is used exclusively to run your application's
-  # test suite.  You never need to work with it otherwise.  Remember that
+  # test suite. You never need to work with it otherwise. Remember that
   # your test database is "scratch space" for the test suite and is wiped
-  # and recreated between test runs.  Don't rely on the data there!
+  # and recreated between test runs. Don't rely on the data there!
   config.cache_classes = true
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  # Do not eager load code on boot. This avoids loading your whole application
+  # just for the purpose of running a single test. If you are using a tool that
+  # preloads Rails for running tests, you may have to set it to true.
+  config.eager_load = false
 
-  # Show full error reports and disable caching
+  # Configure static file server for tests with Cache-Control for performance.
+  config.serve_static_files   = true
+  config.static_cache_control = 'public, max-age=3600'
+
+  # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
-  # Raise exceptions instead of rendering exception templates
+  # Raise exceptions instead of rendering exception templates.
   config.action_dispatch.show_exceptions = false
 
-  # Disable request forgery protection in test environment
-  config.action_controller.allow_forgery_protection    = false
+  # Disable request forgery protection in test environment.
+  config.action_controller.allow_forgery_protection = false
 
   # Tell Action Mailer not to deliver emails to the real world.
   # The :test delivery method accumulates sent emails in the
   # ActionMailer::Base.deliveries array.
   config.action_mailer.delivery_method = :test
 
-  # Use SQL instead of Active Record's schema dumper when creating the test database.
-  # This is necessary if your schema can't be completely dumped by the schema dumper,
-  # like if you have constraints or database-specific column types
-  # config.active_record.schema_format = :sql
+  # Randomize the order test cases are executed.
+  config.active_support.test_order = :random
 
-  # Print deprecation notices to the stderr
+  # Print deprecation notices to the stderr.
   config.active_support.deprecation = :stderr
 
-  # Logger settings
-  config.logger = Logger.new(config.paths.log.first)
-  config.logger.level = Logger::WARN
+  # Raises error for missing translations
+  # config.action_view.raise_on_missing_translations = true
 
-  # Sendmail
+  # Mail delivery method
   config.action_mailer.delivery_method = :smtp
-  config.action_mailer.smtp_settings   = {
-    :address        => 'localhost',
-    :port           => 25,
-    :domain         => 'localhost',
-    :user_name      => nil,
-    :password       => nil,
-    :authentication => nil
-  }
-
-  # IMAP
-  Joruri.config.imap_settings = {
-    :address        => '192.168.0.4',
-    :port           => 143,
-    :usessl         => false
-  }
-
-  # SSO
-  Joruri.config.sso_settings = {
-    :gw => {
-      :host        => 'localhost',
-      :port        => 80,
-      :usessl      => false,
-      :path        => 'api/air_sso'
-    }
-  }
 end
-
