@@ -1,22 +1,11 @@
-# encoding: utf-8
-
 truncate_table "sessions"
-truncate_table "sys_creators"
-truncate_table "sys_editable_groups"
 truncate_table "sys_files"
 truncate_table "sys_groups"
 truncate_table "sys_languages"
 truncate_table "sys_ldap_synchros"
 truncate_table "sys_maintenances"
 truncate_table "sys_messages"
-truncate_table "sys_object_privileges"
-truncate_table "sys_publishers"
-truncate_table "sys_recognitions"
-truncate_table "sys_role_names"
 truncate_table "sys_sequences"
-truncate_table "sys_tasks"
-truncate_table "sys_unids"
+truncate_table "sys_user_logins"
 truncate_table "sys_users"
 truncate_table "sys_users_groups"
-truncate_table "sys_users_roles"
-truncate_table "sys_user_logins"
