@@ -1,7 +1,7 @@
 #!/usr/bin/env ruby
 DONE_FLAG = "/tmp/#{$0}_done"
 
-PASSENGER_VERSION = '3.0.9'
+PASSENGER_VERSION = '5.0.30'
 
 puts '#### Install Apache ####'
 exit if File.exist?(DONE_FLAG)
@@ -17,7 +17,7 @@ end
 def centos
   puts "It's CentOS!"
 
-  system 'yum install -y httpd-devel shared-mime-info'
+  system 'yum install -y httpd-devel'
 
   httpd_conf = '/etc/httpd/conf/httpd.conf'
 
@@ -43,7 +43,7 @@ def centos
   unless File.exist?(passenger_conf)
     system 'yum install -y curl-devel'
     system "gem install passenger -v #{PASSENGER_VERSION}"
-    system 'passenger-install-apache2-module'
+    system 'passenger-install-apache2-module -a'
 
     File.open(passenger_conf, File::RDWR|File::CREAT, 0644) do |f|
       f.flock(File::LOCK_EX)
