@@ -1,4 +1,3 @@
-# encoding: utf-8
 class Sys::Lib::Mail::Attachment
   attr_reader :seqno
   attr_reader :content_type
@@ -6,92 +5,70 @@ class Sys::Lib::Mail::Attachment
   attr_reader :body
   attr_reader :size
   attr_reader :transfer_encoding
-  
+
   def initialize(attributes = {})
     attributes.each {|name, value| eval("@#{name} = value")}
   end
-  
+
   def image?
-    @content_type =~ /^image/i ? true : false
+    !(@content_type =~ /^image/i).nil?
   end
-  
-  def image_width
+
+  def disposition
+    image? ? 'inline' : 'attachment'
+  end
+
+  def thumbnail(width: 64, height: 64, format: 'jpeg', quality: 50, method: :thumbnail)
     return nil unless image?
-    begin
-      require 'RMagick'
-      image = Magick::Image.from_blob(data).shift
-      if image.format =~ /(GIF|JPEG|PNG)/
-        @image_width = image.columns
-        @image_height= image.rows
+
+    image = Magick::Image.from_blob(@body).shift
+    thumb_w, thumb_h = self.class.reduce_size(image.columns, image.rows, width, height)
+    if image.columns > thumb_w || image.rows > thumb_h
+      image.format = format.to_s
+      case method
+      when :sample
+        image.sample!(thumb_w, thumb_h)
+      when :thumbnail
+        image.thumbnail!(thumb_w, thumb_h)
+      else
+        image.resize!(thumb_w, thumb_h)
       end
-    rescue => e
-    end
-    @image_width
-  end
-  
-  def image_height
-    return nil unless image_width
-    @image_height
-  end
-  
-  def reduced_size(options = {})
-    return nil unless image?
-    src_w  = image_width.to_f
-    src_h  = image_height.to_f
-    dst_w  = options[:width].to_f
-    dst_h  = options[:height].to_f
-    src_r    = (src_w / src_h)
-    dst_r    = (dst_w / dst_h)
-    if dst_r > src_r
-      dst_w = (dst_h * src_r);
+      image.to_blob { self.quality = quality }
     else
-      dst_h = (dst_w / src_r);
+      image.to_blob
     end
-    
-    if options[:css]
-      return "width: #{dst_w.ceil}px; height:#{dst_h.ceil}px;"
-    end
-    return {:width => dst_w.ceil, :height => dst_h.ceil}
+  rescue => e
+    error_log(e)
+    nil
   end
-  
-  def thumbnail(options)
-    begin
-      require 'RMagick'
-      image = Magick::Image.from_blob(@body).shift
-      raise 'NotImage' unless image.format =~ /(GIF|JPEG|PNG|BMP)/
-      src_w = image.columns.to_f
-      src_h = image.rows.to_f
-      dst_w = options[:width].to_f
-      dst_h = options[:height].to_f
-      if src_w > dst_w || src_h > dst_h
-        src_r = (src_w / src_h)
-        dst_r = (dst_w / dst_h)
-        if dst_r > src_r
-          dst_w = (dst_h * src_r);
-        else
-          dst_h = (dst_w / src_r);
-        end
-        image.format = options[:format].to_s if options[:format]
-        image.thumbnail!(dst_w.ceil, dst_h.ceil)
-      end
-      return image.to_blob { self.quality = options[:quality] if options[:quality] }
-    rescue => e
-      return nil
-    end
-  end
-  
+
   def css_class
-    if ext = File::extname(@name).downcase[1..5]
-      ext = ext.gsub(/[^0-9a-z]/, '')
-      return 'iconFile icon' + ext.gsub(/\b\w/) {|word| word.upcase}
+    if ext = File::extname(@name.to_s).downcase[1..5]
+      ext = ext.gsub(/[^0-9a-z]/, '').gsub(/\b\w/) { |word| word.upcase }
+      "iconFile icon#{ext}"
+    else
+      'iconFile'
     end
-    return 'iconFile'
-  rescue
-    return 'iconFile'
   end
-  
+
   def eng_unit
-    return @eng_unit if @eng_unit
-    @eng_unit = "#{Util::Unit.eng_unit(@size, "Bytes")}"
+    @eng_unit ||= ApplicationController.helpers.number_to_human_size(@size, precision: 0, locale: :en)
+  end
+
+  class << self
+    def reduce_size(src_width, src_height, dst_width, dst_height)
+      src_w = src_width.to_f
+      src_h = src_height.to_f
+      dst_w = dst_width.to_f
+      dst_h = dst_height.to_f
+      src_r = src_w / src_h
+      dst_r = dst_w / dst_h
+      if dst_r > src_r
+        dst_w = dst_h * src_r
+      else
+        dst_h = dst_w / src_r
+      end
+      return dst_w.ceil, dst_h.ceil
+    end
   end
 end
