@@ -1,12 +1,11 @@
 module Sys::Model::Auth::Free
-  def readable
-    return self
+  extend ActiveSupport::Concern
+
+  included do
+    scope :readable, -> { all }
+    scope :editable, -> { all }
   end
-  
-  def editable
-    return self
-  end
-  
+ 
   def creatable?
     true
   end
