@@ -1,190 +1,177 @@
-// require prototype.js, cookiemanager.js
-
 /**
  * Navigation
  */
-function Navigation(settings) {
-  var self = this;
-  this.settings = settings;
-  
-  this.onLoad = function() {
-    Navigation.theme();
-    Navigation.fontSize();
-    Navigation.ruby($(this.settings['ruby']), null, $(this.settings['notice']));
-    
-    if (this.settings['theme']) {
-      for (var k in this.settings['theme']) {
-        Event.observe($(k), 'click', function(evt) {self.theme(evt)}, false);
-      }
-    }
-    if (this.settings['fontSize']) {
-      for (var k in this.settings['fontSize']) {
-        Event.observe($(k), 'click', function(evt) {self.fontSize(evt)}, false);
-      }
-    }
-    if (this.settings['ruby']) {
-      var k = this.settings['ruby'];
-      if (k) {
-        Event.observe($(k), 'click', function(evt) {self.ruby(evt)}, false);
-      }
-    }
-    if (this.settings['talk']) {
-      var k = this.settings['talk'];
-      if (k) {
-        $(this.settings['talk']).addClassName('talkOff');
-        Event.observe($(k), 'click', function(evt) {self.talk(evt)}, false);
-      }
-    }
-  }
-  Event.observe(window, 'load', function(){ self.onLoad() }, false);
-  
-  this.theme = function(evt) {
-    var element = Event.element(evt);
-    if (this.settings['theme'][element.id]) {
-      Navigation.theme(this.settings['theme'][element.id]);
-    }
-  }
-  
-  this.fontSize = function(evt) {
-    var element = Event.element(evt);
-    if (this.settings['fontSize'][element.id]) {
-      Navigation.fontSize(this.settings['fontSize'][element.id]);
-    }
-  }
-  
-  this.ruby = function(evt) {
-    var element = Event.element(evt);
-    if ($(element).className.match(/(^| )rubyOn( |$)/)) {
-      Navigation.ruby(element, 'off', $(this.settings['notice']));
-    } else {
-      Navigation.ruby(element, 'on', $(this.settings['notice']));
-    }
-  }
-  
-  this.talk = function(evt) {
-    var element = Event.element(evt);
-    Navigation.talk(element, $(this.settings['player']), $(this.settings['notice']));
-  }
+function Navigation() {
 }
 
-/**
- * Changes the stylesheets.
- */
+function Navigation_initialize(settings) {
+  Navigation.settings = settings
+  
+  if (Navigation.settings['theme']) {
+    jQuery.each(Navigation.settings['theme'], function(key, val) {
+      $(key).on('click', function(){
+        Navigation.theme(val);
+        return false;
+      });
+    });
+  }
+  
+  if (Navigation.settings['fontSize']) {
+    jQuery.each(Navigation.settings['fontSize'], function(key, val) {
+      $(key).on('click', function(){
+        Navigation.fontSize(val);
+        return false;
+      });
+    });
+  }
+  
+  if (Navigation.settings['ruby']) {
+    $(Navigation.settings['ruby']).on('click', function(){
+      var flag = ($(this).attr('class') + '').match(/(^| )rubyOn( |$)/);
+      Navigation.ruby( (flag ? 'off' : 'on') );
+      return false;
+    });
+  }
+  
+  if (Navigation.settings['talk']) {
+    $(Navigation.settings['talk']).on('click', function(){
+      var flag = ($(this).attr('class') + '').match(/(^| )talkOn( |$)/);
+      Navigation.talk( (flag ? 'off' : 'on') );
+      return false;
+    });
+  }
+  
+  Navigation.theme();
+  Navigation.fontSize();
+  Navigation.ruby();
+}
+Navigation.initialize = Navigation_initialize;
+  
+    // if (this.settings['talk']) {
+      // var k = this.settings['talk'];
+      // if (k) {
+        // $(this.settings['talk']).addClassName('talkOff');
+        // Event.observe($(k), 'click', function(evt) {self.talk(evt); Event.stop(evt);}, false);
+      // }
+    // }
+  
+  // this.talk = function(evt) {
+    // var element = Event.element(evt);
+    // Navigation.talk(element, $(this.settings['player']), $(this.settings['notice']));
+  // }
+
 function Navigation_theme(theme) {
   if (theme) {
-    (new CookieManager()).setCookie('navigation_theme', theme);
+    $.cookie('navigation_theme', theme, {path: '/'});
   } else {
-    theme = (new CookieManager()).getCookie('navigation_theme');
+    theme = $.cookie('navigation_theme');
     if (!theme) return false;
   }
-  var links = document.getElementsByTagName('link');
-  for (var i = 0; i < links.length; i++) {
-    if (links.item(i).title != '') links.item(i).disabled = true;
-  }
-  for (var i = 0; i < links.length; i++) {
-    if (links.item(i).title == theme) links.item(i).disabled = false;
-  }
-  return false;
+  $('link[title]').each(function() {
+    this.disabled = true;
+    if (theme == $(this).attr('title')) this.disabled = false;
+  });
 }
 Navigation.theme = Navigation_theme;
 
-/**
- * Changes the fontsize.
- */
 function Navigation_fontSize(size) {
   if (size) {
-    (new CookieManager()).setCookie('navigation_font_size', size);
+    $.cookie('navigation_font_size', size, {path: '/'});
   } else {
-    size = (new CookieManager()).getCookie('navigation_font_size');
+    size = $.cookie('navigation_font_size');
     if (!size) return false;
   }
-  //document.body.style.fontSize = size;
-  if ($('contentBody')) $('contentBody').setStyle({ fontSize: size });
-  return false;
+  $('#contentBody').css('font-size', size);
 }
 Navigation.fontSize = Navigation_fontSize;
 
-/**
- * Toggles the Ruby.
- */
-function Navigation_ruby(element, flag, noticeContainer) {
-  if (!element) return false;
+function Navigation_ruby(flag) {
+  if (!Navigation.settings['ruby']) return false;
+  var elem = $(Navigation.settings['ruby']);
+  if (flag == 'on') {
+    $.cookie('navigation_ruby', flag, {path: '/'});
+    if (location.pathname.search(/\/$/i) != -1) {
+      location.href = location.pathname + "index.html.r" + location.search;
+    } else if (location.pathname.search(/\.html\.mp3$/i) != -1) {
+      location.href = location.pathname.replace(/\.html\.mp3$/, ".html.r") + location.search;
+    } else if (location.pathname.search(/\.html$/i) != -1) {
+      location.href = location.pathname.replace(/\.html$/, ".html.r") + location.search;
+    } else if (location.pathname.search(/\.html$/i) != -1) {
+      location.href = location.pathname.replace(/\.html$/, ".html.r") + location.search;
+    } else {
+      location.href = location.href.replace(/#.*/, '');
+    }
+  } else if (flag == 'off') {
+    $.cookie('navigation_ruby', flag, {path: '/'});
+    if (location.pathname.search(/\.html\.r$/i) != -1) {
+      location.href = location.pathname.replace(/\.html\.r$/, ".html") + location.search;
+    } else {
+      location.reload();
+    }
+  }
+  if (flag) return;
   
-  if (flag) {
-    //change
-    (new CookieManager()).setCookie('navigation_ruby', flag);
-    
-    //redirect
-    if (flag == 'on') {
-      if (location.pathname.search(/\.html$/i) != -1) {
-        location.href = location.pathname.replace(/\.html/, ".html.r") + location.search;
-      } else if (location.pathname.search(/\.html\.mp3$/i) != -1) {
-        location.href = location.pathname.replace(/\.html\.mp3/, ".html.r") + location.search;
-      } else if (location.pathname.search(/\/$/i) != -1) {
-        location.href = location.pathname + "index.html.r" + location.search;
-      }
+  if ($.cookie('navigation_ruby') == 'on') {
+    if (location.pathname.search(/\/$/i) != -1) {
+      location.href = location.pathname + "index.html.r" + location.search;
+    } else if (location.pathname.search(/\.html$/i) != -1) {
+      location.href = location.pathname.replace(/\.html/, ".html.r") + location.search;
     } else {
-      if (location.pathname.search(/\.html\.r/i) != -1) {
-        location.href = location.pathname.replace(/\.html\.r/,".html") + location.search ;
-      }
+      elem.removeClass('rubyOff');
+      elem.addClass('rubyOn');
+      Navigation.notice();
     }
-    
   } else {
-    // render
-    if ((new CookieManager()).getCookie('navigation_ruby') == 'on') {
-      element.removeClassName('rubyOff');
-      element.addClassName('rubyOn');
-      Navigation.showNotice(noticeContainer);
-    } else {
-      element.removeClassName('rubyOn');
-      element.addClassName('rubyOff');
-    }
+    elem.removeClass('rubyOn');
+    elem.addClass('rubyOff');
   }
 }
 Navigation.ruby = Navigation_ruby;
 
-/**
- * Navigation Speaker.
- */
-function Navigation_Talk(element, player, container) {
-  Navigation.showNotice(container);
+function Navigation_Talk(flag) {
+  var player = $(Navigation.settings['player']);
+  var elem   = $(Navigation.settings['talk']);
+  if (!player || !elem) return false;
   
-  if (element.className.match(/(^| )talkOn( |$)/)) {
-    element.removeClassName('talkOn');
-    element.addClassName('talkOff');
+  Navigation.notice();
+  
+  if (flag == 'off') {
+    elem.removeClass('talkOn');
+    elem.addClass('talkOff');
   } else {
-    element.removeClassName('talkOff');
-    element.addClassName('talkOn');
+    elem.removeClass('talkOff');
+    elem.addClass('talkOn');
   }
-  
+   
   var uri = location.pathname;
-  if (uri.match(/\/$/)) {
-    uri += 'index.html';
-  }
+  if (uri.match(/\/$/)) uri += 'index.html';
+  uri = uri.replace(/\.html\.r$/, '.html');
   
   var now   = new Date();
-  var param = '?' + now.getDay() + now.getHours();
+  var param = '?85' + now.getDay() + now.getHours();
   
   if (player) {
     uri += '.mp3' + param;
-    if (player.innerHTML == '') {
-      html = '<script type="text/javascript" src="/_common/swf/niftyplayer/niftyplayer.js"></script>' +
-      '<object classid="clsid:D27CDB6E-AE6D-11cf-96B8-444553540000"' +
-      ' codebase="http://download.macromedia.com/pub/shockwave/cabs/flash/swflash.cab#version=6,0,0,0"' +
-      ' width="165" height="37" id="niftyPlayer1" align="">' +
-      '<param name=movie value="/_common/swf/niftyplayer/niftyplayer.swf?file=' + uri + '&as=1">' +
-      '<param name=quality value=high>' +
-      '<param name=bgcolor value=#FFFFFF>' +
-      '<embed src="/_common/swf/niftyplayer/niftyplayer.swf?file=' + uri + '&as=1" quality=high bgcolor=#FFFFFF' +
-      ' width="165" height="37" name="niftyPlayer1" align="" type="application/x-shockwave-flash"' +
-      ' swLiveConnect="true" pluginspage="http://www.macromedia.com/go/getflashplayer">' +
-      '</embed>' +
+    if (player.html() == '') {
+      html = '<div id="navigationTalkCreatingFileNotice" style="display: none;">ただいま音声ファイルを作成しています。しばらくお待ちください。</div>';
+      html += '<object type="application/x-shockwave-flash" data="/_common/swf/dewplayer/dewplayer-vol.swf" width="240" height="20" id="dewplayer" name="dewplayer">' +
+      '<param name="wmode" value="transparent" />' +
+      '<param name="movie" value="/_common/swf/dewplayer/dewplayer-vol.swf" />' +
+      '<param name="flashvars" value="mp3=' + uri + '&autostart=1" />' +
       '</object>';
-      player.innerHTML = html;
+      player.html(html);
+
+      $.ajax({type: "HEAD", url: uri, data: {file_check: '1'}, success: function(data, status, xhr) {
+        var type = xhr.getResponseHeader('Content-Type');
+        if (type.match(/^audio/)) {
+          $('#navigationTalkCreatingFileNotice').hide();
+        } else { 
+          $('#navigationTalkCreatingFileNotice').show();
+        }
+      }});
     } else {
-      uri += '.m3u' + param;
-      player.innerHTML = '';
-      $('navigationNotice').remove();
+      player.html('');
+      if ($.cookie('navigation_ruby') != 'on') Navigation.notice('off');
     }
   } else {
     location.href = uri;
@@ -192,18 +179,22 @@ function Navigation_Talk(element, player, container) {
 }
 Navigation.talk = Navigation_Talk;
 
-/**
- * Shows the notice.
- */
-function Navigation_showNotice(container) {
-  var notice = $('navigationNotice');
-  if (!container || notice) {
-    return true;
+function Navigation_notice(flag) {
+  var wrap   = Navigation.settings['notice'] || 'container';
+  var notice = $('#navigationNotice');
+  
+  if (flag == 'off') {
+    notice.remove();
+    return false;
   }
-  var notice = document.createElement('div'); 
+  if (notice.size()) return false;
+  
+  var elem = $(Navigation.settings['notice']);
+  notice = document.createElement('div'); 
   notice.id = 'navigationNotice'; 
-  notice.innerHTML = 'ふりがなと読み上げ音声は、システムにより自動的に生成しています。' +
-    '人名、地名、用語などが、正確に発音されない場合がありますので、ご了承ください。';
-  container.insertBefore(notice, container.firstChild);
+  notice.innerHTML = 'ふりがなと読み上げ音声は，' +
+    '人名，地名，用語等が正確に発音されない場合があります。';
+  // $(wrap + ' *:first').before(notice);
+  $('#accessibilityTool').prepend(notice);
 }
-Navigation.showNotice = Navigation_showNotice;
+Navigation.notice = Navigation_notice;
